requirements.txt
paho-mqtt>=1.4.0
boto3>=1.9.218